requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.2

# Vector databases - Compatible versions
chromadb==0.4.15
faiss-cpu==1.7.4

# LLM & Embeddings - Compatible versions
openai==1.3.0
tiktoken==0.5.2
sentence-transformers==2.2.2

# Document processing
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
pypdf==3.17.4
python-docx==1.1.0

# UI
streamlit==1.29.0
plotly==5.18.0

# Utilities
tqdm==4.66.1
pytest==7.4.3